requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
datasets==4.2.0
dill==0.4.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
multidict==6.7.0
multiprocess==0.70.16
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pip==25.2
propcache==0.4.1
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
